services/inference/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
psycopg2-binary==2.9.9
scikit-learn==1.4.0
numpy==1.26.3
python-dotenv==1.0.0
